Changed directory structure to follow golang recomendations
