+++ src/fileio/reader/reader.go
@@ -1,0 +1,1 @@
+// TODO: Make a generic reader for the image file and config file
